Prototyped the zWait and zSave routines (WIP)
I've started working on a zWait dialog to show a spinner while
something is working on something else (such as a file load, save,
upload or download to the EPROM). I need a way to start the zWait
routine, run operation x() and then end zWait.

Also started working with bincopy, which manages EPROM HEX file
formats. I've created a simple routine to save the existing binary
data in the stow[] to a file called 'junk.ihx', in Intel hex format.
Still need to work on saving as s19, ihx or binary based on the file
ending (.s19, .ihx or .bin'). Also need to work on the loading of
these file formats. At the moment the Tektronic and Mostek formats are
not supported. I doubt I'll add them.