zSave is working with the extensions. zLoad is still a work in progress.
I have zSave working with .srec, .s19., ihx and .bin. I haven't tried
.hex or .bin with bincopy (I'm using a write instead). I've also done
a bit of work with zLoad but it's not actually reading any files in
yet.

There's absolutely no error checking at this moment and that needs a
lot of thought. How do we handle a failure on read or write of a file?
Just blowing up is not a good idea.