Returned to working on loading ROM images
Fixed a multituide of small defects. Still need to do a lot more error checing.
Added a skeleton for a tools section where I can manipulate various parameter.
I've added a way to bypass the initial check for the device (hit s). I've fixed
a bug with reading and writing files (need the full path and filename). Fixed a
bug where reading past the end of stow caused a crash. Still need to move the
code to a curses wrapper so crashes get caught and the program does a stty
sane. Need to add a status/diag box and I wish I could figure out how to add a
new thread as needed to do things like the spinner and other theings where we
wait on completion of some task.