Started work on editor, worked on loaders and refactored the code
Did an amazinly stupid thing. I made a lot of changes without comming
them to git. Fortunately I didn't screw up too badly and was able to
recover a change that messed up everything. I'll now commit the
changes for the loaders (which still needs work), the editor (which is
a thing in progress) and the refactored code.